....I..PS. [ZBX-4941] add default DBHost configuration parameter value "localhost"
If DBHost parameter is missing in configuration it will be set to
"localhost". Empty value is allowed, in this case the connection
method (Unix domain socket, TCP etc.) is based on used backend
library. E. g. in case of PostgreSQL if you would like to use Unix
domain socket to connect to database DBHost should be set to empty
string. Missing parameter will result in TCP connection to
"localhost". In case of MySQL both empty string and missing parameter
will result in attempt to use Unix domain socket.


git-svn-id: svn://svn.zabbix.com/trunk@27417 97f52cf1-0a1b-0410-bd0e-c28be96e8082