...GI..... [ZBX-5171] fix zabbix_agent compilation errors on HP-UX
In Zabbix 2.0 net.if.discovery functionality was introduced. With this
functionality in zabbix_agent for HP-UX 2 new function calls were
added:

  if_nameindex()
  if_freenameindex()

From the man page it is stated that in order to use these functions
application should be linked with libipv6 library. This commit changes
configure.in to include -lipv6 to AGENT_LIBS variable if detected os
is HP-UX.


git-svn-id: svn://svn.zabbix.com/branches/dev@28436 f4e7fd02-d210-4e34-a1e0-d4093fb66cc2