....I..... [ZBX-5386] allow unixODBC support without odbc_config
Some GNU/Linux distributions have unixODBC packages without
odbc_config binary, that you can use to get things needed to
add support for unixODBC to your application.

Before this fix when building Zabbix server with unixODBC support if
odbc_config binary was missing configure would result in an error.
This commit changes it so that if odbc_config is missing we continue
without additional CFLAGS and LDFLAGS with just additional -lodbc. On
a Debian 6.0 this allowed successful compilation.


git-svn-id: svn://svn.zabbix.com/branches/dev@29541 f4e7fd02-d210-4e34-a1e0-d4093fb66cc2