........S. [ZBX-5386] fix SQL query result handling
Before this fix SQL queries of "Database monitor" items where not
properly checked for:

  - empty result set
  - NULL values

In both cases an item became UNSUPPORTED (except for the null value
for "Character" value type, which isn't probably correct) but the
error message was provided only in case NULL value was received for
Integer value type. An error message was:

  item [db.odbc.select[...]] became not supported: Received value [@?]
  is not suitable for value type [Numeric (unsigned)] and data type
  [Decimal]

The @? is garbage.

This commit adds handling of such situations. In both cases an item
becomes unsupported no matter what value type an item has. The error
messages are respectively:

  - SQL query returned empty result.
  - SQL query returned NULL value.


git-svn-id: svn://svn.zabbix.com/branches/dev@29542 f4e7fd02-d210-4e34-a1e0-d4093fb66cc2