...GI..... [ZBX-5171] get rid of agent ipv6 library dependency on older HP-UX machines
It was decided to get rid of ipv6 library dependency on the agent
side. This was needed for LLD. If the machine is older than 11.31
different code is used in order to get the list of available network
interfaces when "net.if.discovery" is used.

Before this fix:
  - agent required ipv6 library on HP-UX prior to 11.31 in order for
    low level network discovery to work

After this fix:
  - agent does not require ipv6 library on HP-UX prior to 11.31 in
    order for low level network discovery to work


git-svn-id: svn://svn.zabbix.com/trunk@29953 97f52cf1-0a1b-0410-bd0e-c28be96e8082