...G...... [ZBX-5571] fixed reporting of system.sw.packages for multiple package systems ...G...... [ZBX-5559] fixed reporting of system.sw.packages with short package listing
When using zbx_vector_str_create() and zbx_vector_str_append(), it is important to not forget
deallocate memory used for string vector elements before calling zbx_vector_str_destroy().
Before ZBX-4802 the memory deallocation was done in print_packages() function.
Trying to improve code clarity, the ZBX-4802 was registered but it was solved incorrectly.
The new solution for ZBX-5571 puts the memory deallocation in the same function where it is allocated.

This change also solves ZBX-5559 as the fix is trivial and in the same file as for ZBX-5571.


git-svn-id: svn://svn.zabbix.com/branches/dev@30218 f4e7fd02-d210-4e34-a1e0-d4093fb66cc2