.......PS. [ZBX-3379] add support for SourceIP when doing snmp checks
On some systems snmp checks were failing if SourceIP parameter was
specified in server or proxy configuration. This happened because
local host was specified without local port when opening snmp
session. The error message was:

  "Cannot open snmp session"

When snmp session is opened if you want to specify the source IP the
"localname" member of "struct snmp_session" must be set.

Before this change:
  - "localname" was set to the value of SourceIP

After this change:
  - "localname" is set to "host:0" where "host" is the value of SourceIP


git-svn-id: svn://svn.zabbix.com/branches/dev@31977 f4e7fd02-d210-4e34-a1e0-d4093fb66cc2