....I..... [ZBX-6637] added -m64 compilation flag for solaris to be able to understand all attributes of a 64-bit processes
If 32-bit zabbix agent is runing on 64-bit Solaris host, proc.mem returns 0 for some processes even if they use memory. 
This happens because existing interfaces and data structures that describe the process are not large enough to contain the 64-bit quantities.
In order to avoid that the application has to be recompiled as 64-bit program in order to work with both 32-bit processes and 64-bit processes.


git-svn-id: svn://svn.zabbix.com/branches/dev@36473 f4e7fd02-d210-4e34-a1e0-d4093fb66cc2