....I..... [RSM-1] 121: add option --continue: continue from the timestamp of last updated api data
Continue processing API data from the timestamp of the last run with
--continue. In case of first run with --continue all available data
will be processed. The continue token is saved per each TLD-service
pair separately.

Note, that continue token will not be updated if this option was
specified together with --dry-run.

This option cannot be used together with option --period.


git-svn-id: svn://svn.zabbix.com/branches/2.0.rsm@52560 f4e7fd02-d210-4e34-a1e0-d4093fb66cc2