.......PS. [ZBX-9436] zabbix_server should not exit if database re-connection occurs during active transaction
During active transaction if a reconnect occurs the transaction is
marked as failed and all subsequent queries are ignored. On the other
hand all the queries during a connection to the database are allowed.
This commit introduces new variable txn_reconn which indicates if
re-connection occurs during a transaction.

If connection to the database is lost database engine rolls back (or
at least should do it) all active transactions.


git-svn-id: svn://svn.zabbix.com/branches/dev@52950 f4e7fd02-d210-4e34-a1e0-d4093fb66cc2