........S. [ZBX-11805] Fixed the aggregation checks log output
For fixing the log output i needed to modify the calles related on how to
compose the log strings. In particular, the common call was removed
and the two situation where we need aggregate_quote_groups() now
have a dedicated call.
This is due to the fact that the new message for indicating that no groups
are correct, need to be formatted in three different steps, where the
groups are in the middle.