.......... [DEV-575] Suppress variable scope warnings
As we use old style variable declarations on top of the function scope,
cppcheck will complain a lot about those variable if we scan with
standard as C99.
But we respect that standard in all the other coding guidelines, so just
suppress it for now until we will want to do something about it.