...G...... [ZBX-12101] Improved log initialization
Fix the breakage that occured in trunk regarding the zabbix_sender.
Explicitly check if the type is differnt than LOG_TYPE_UNDEFINED as
the last condition in the if..else chain in zabbix_open_log.
If the condition is true, we return FAIL. Else, we write logs on the
stderr as usual.
In this way we can cover with proper failure if a log type value is
something we really don't know.