...G...... [ZBXNEXT-3910] Make the agent monitor additional processes states
Now the agent will be able to monitor processes that are in uninterruptible state
(generally disk I/O) or that are in trace state (if a debugger is attached to a
process for example).
Add this support for Linux, FreeBSD, OpenBSD and NetBSD. AIX, Solaris and
Windows are not included.


git-svn-id: svn://svn.zabbix.com/trunk@70560 97f52cf1-0a1b-0410-bd0e-c28be96e8082