...G...... [ZBX-12397] fixed system.cpu.num for 64-bit Windows for Windows 7 and later
This fix allow the zabbix agent to report the correct number of logical CPUs
if it's running on WIndows 7 or later. The process group feature introduced
in Windows 7, allowed, by default, a process to see only up to 64 CPUs. Now
It's possible to monitor and report poperly CPU counts greater than 64.
The fix is also compatible with previous versions of Windows that doesn't
have this capability.


git-svn-id: svn://svn.zabbix.com/trunk@71102 97f52cf1-0a1b-0410-bd0e-c28be96e8082