...G...... [ZBX-12653] fixed heap corruption in Windows agent
The _snwprintf_s take, as second argument, the size in "words" of the
buffer passed as first argument NOT a size in bytes. So, just pass the
variable named "size" that already contain it and also modify the third
parameter for telling the function to write maximum "size - 1" wide
characters.

Patch from Ronnie Kaech. Thanks!

[svn merge -c r72371 ^/branches/3.4]


git-svn-id: svn://svn.zabbix.com/trunk@72372 97f52cf1-0a1b-0410-bd0e-c28be96e8082