....I..... [ZBX-12793] fixed compilers checks during ./configure
Add to the main configuration script the explicit checks for cc
and c99. The cc utility is, in theory, deprecated, but it's much
more common to find than c99 in supported systems.
So add a check for both the standard commands (cc first, c99
second) and after that for the two specific compilers we test
(gcc and clang).
No changes on the usage of the CC and CFLAGS environment
variables.