........S. [ZBXNEXT-4108] splitted all the updated patches.
While testing the database update, It was clear that it's way
better to have every upgrade step to be by itself, even for
ones that can share the same queries or default values.
Debugging can be difficult without this and a proper feedback
from a user, if required, can be problematic.
We were thinking to wait for Sasha's opinion on that, but I
was convinced by the idea and committed it.

Discussed with Andris Z.