........S. [ZBXNEXT-3480] moved zbx_inventory_value_t to dbcache.h
The zbx_inventory_value_t was moved in dbcache.h for being
used also in other files.
The value_esc field is now named as value. The change is in
preparation to the next diff that will take care of escaping
the value only when it will be added to the sql statement.
This is necessary for being able to update an host inventory
field in the configuration cache on the fly.