........S. [ZBXNEXT-3480] remove the new lock
As vso pointed out, it's actually impossible to use a
separated lock for the new cache.
Too many potential race conditions and also a good
chance of deadlocking.

For now, use for the new cache the global configuration
cache lock until we will find a way for solving this
potential problem.