..F.....S. [ZBXNEXT-4108] implemented search of problems by trigger name
It will now be possible to search problems by the related
trigger description in the zabbix front-end. This is done
through a new "name" field in the events and problem
tables.
Macros used in the event or problem name will be
expanded in the same way as it's done currently for the
trigger description.


git-svn-id: svn://svn.zabbix.com/trunk@75329 97f52cf1-0a1b-0410-bd0e-c28be96e8082