..F.....S. [ZBXNEXT-4002] implemented elasticsearch as history storage
Elasticsearch can be used as a history storage backend
now for storing all the type of data supported by zabbix.
The feature can be enabled on the server through the
two new configuration parameters HistoryStorageURL
(the elasticsearch HTTP path) and HistoryStorageTypes
(the parameter to decide which data types must be
stored in elasticsearch).

The frontend also support to retrieve data from elasticsearch
using the $HISTORY['url'] and $HISTORY['types'] parameters.

[svn merge -c 75184 ^/branches/3.4]


git-svn-id: svn://svn.zabbix.com/trunk@75502 97f52cf1-0a1b-0410-bd0e-c28be96e8082