........S. [ZBXNEXT-4002] add back proper return assignement
The assignement for the ret variable was dropped during
merge. Addit back as it is required in a next step for
deciding to prepare or not the history for modules.

Also, fix a bug where we check for ZBX_DB_OK instead
of SUCCEED before calling DCmodule_prepare_history()
and friends.

Spotted by coverity