Ensure the testPageDashboard runs successfully
Sometimes when the Filter button was clicked, the filter form was not visabled. And at this
time when we query the 'zbx_filter' by name, we could not find any valid zbx_filter, and then
the case failed.
In order to ensure that the case run successfully, we should check the filter form is whether visable
or not, if not, just click the Filter button again.