...G...PS. [ZBX-20864] initialize result variable before use
In the case of a POLYNOMIAL fit, we used to feed the result value to
zbx_polynomial_timeleft() uninitialized, where it could end up having an
arbitrary value after some arithmetic. Initialize it to a value outside
the lower bound so we get an error value instead of an arbitrary,
potentially valid, value.