...G...... [ZBX-21373] counted proc args on openbsd manually instead of using a syscall
It appears that data integrity is not maintained between two calls to
sysctl() asking for kern.procargs, when first asking for argv itself and
then asking for nargv. The latter call would return a value that was +1
of what was expected, resulting in us trying to collect_args() with an
incorrect argc parameter and derefing the list terminating NULL pointer.

This fixes the issue by counting the args in proc_argv manually and
avoids an extra syscall, which is a win.