.......PS. [ZBXNEXT-7931] changed interface availability tolerance
When interfaces are (re)assigned to proxies, we provide a grace period
before resetting their availability status so as to not immediately mark
them unavailable in case the proxy hasn't exchanged data with the server
since the (re)assignment.

As heartbeats are going away, it seems odd to rely on an upper bound of
the proxies heartbeat frequency parameter to decide the tolerance
interval, so be explicit about it instead. The previous interval
was an hour + 1 minute. As it is expected that servers and proxies
exchange data much more frequently, lower the interval to 10 minutes.