return items and use forEach to iterate over nodes
We know that /api/v1/nodes has a response with the nodes under items,
and we have no further interest in any other data from this request
outside the getNodes function.

Iterate over the nodes using forEach, this makes it easier to handle
the scope of the node you're manipulating.