....I..... [ZBX-22307] made clean rules depend on clean-recursive
Not ideal, as this should be sort of implied (i.e. the target be
recursively built in any declared SUBDIRS), but it just doesn't.
This same rule dependency is used in the root and src/libs Makefile
templates for this same purpose, so just replicate that.

Also declare the clean targets to be .PHONY deps.