...G...... [ZBXNEXT-7569] made system_sw_packages{,_get} use strtok_r
The issue came up when trying to use strtok() inside of any of the
parser functions. As strtok() uses an internal buffer, the parsers use
would result in overwriting that internal buffer and on the following
iteration, the manager functions would end up with an empty buffer and
thus lead to preemptively stop processing the response output.

We already rely on strtok_r() in the server and preprocessing library.