netbsd/net.c: convert to using the now-preferred sysctl() method.
The sysctl() method of getting at the network statistics is the
preferred method on NetBSD nowadays.  The code to do so is lifted
from NetBSD's netstat/if.c and adapted as required.

This fixes a build problem on netbsd-10 where the internal data
structure has changed to use per-CPU counters, and the code has
been tested on NetBSD 8.2, so this method also works in now quite
old NetBSD hosts.