.......... [DEV-2554] added targets for generating SBOMs for the PHP frontend
The convenience target is 'sbom' which depends on the 'sbom.json'
target. A sbom.xml target is also available for generating an SBOM in
XML format.

Both targets explicitly specify the 1.4 version of the CycloneDX SBOM
specification.

The output SBOMs have so far been correct and exhaustive in testing.

The component name defaults to "_root_", and version "dev-master" if not
specified via the --mc-version flag. The flag does not affect the name,
just the version. Ideally, we would want to have the name and version
specified in the composer.json file.