........S. [ZBXNEXT-8593] shifted validator type values down by 1
Drop the reservation of the '0' value and shift the values down by 1.
The original intent was for the default value in the database to be
invalid so old script entries, after the table upgrade, don't get an
arbitrary value for this setting, but that now seems needlessly
cumbersome. Just use the value.