..F....... [ZBX-23066] removed LICENSE copy from the ui directory
The composer.json file specifies the license via an SPDX identifier, so
a specific file is unnecessary. For the time being, it doesn't look like
we'll publish the ui as a package to the composer repositories, so its
not critical to have a copy of the license in the directory.