..F....... [ZBXNEXT-8121] fixed script validation rule building
In the case when a script accepts additional manual input, we add rules
that require some of these fields to be both provided and not empty.
This is done via a 'flags' member in the field description, which is a
bitmask and not a list. Setting it as a list would make the
ApiInputValidator crash due the difference in the expected and real type
of hte field.