........S. [ZBXNEXT-8593] included missing regex and string headers
The recently added local `validate_manualinput()` function calls
`zbx_regexp_match()` and `zbx_str_in_list`, which are provided by the
zbxregexp and zbxstr libraries respectively.

Unfortunately, due to some zbxcommon.h shenanigans, the contents of
those two headers would implicitly become available to this module,
making the code incorrect, yet still capable of compiling. This is a
known issue due to the proliferation of zbxcommon.h includes, which is
not something to be fixed in this set of changes.

Include the headers because it's the right thing to do anyway.