........S. [ZBXNEXT-8593] changed the type of a position tracking variable
Although this is incorrect, we'll abide by the declared type of the
function prototype. In reality, the type of `pos` should be `size_t`, as
(1) the supplied string can be of arbitrary length (but usually limited
to MAX_STRING_LEN as defined by zbxcommon.h), and (2) the current type
implies that the function can backtrack into memory prior to where the
string itself starts.