....I..... [ZBX-22307] fixed database patch template targets
The build targets are only necessary when we generate the schemas and
patches. When a source dist is produced and then used to configure a
build, these files are already generated and building a schema isn't
required.

When make is invoked with the default or `all` target, if these file
generating targets are available, they end up being built, which fails,
as we don't ship the generating scripts in the srcdist. This is why they
should be wrapped in the `DBSCHEMA` conditional to avoid these failures.

Nasty. This fixes it.