.......... [DEV-2707] broke up go module SBOM target into multiple ones
Previously, we generated a single SBOM for the entire modules. However,
the module also carries other applications, such as the web reporting
service (zabbix_web_service) and the mock server (mock_server) for
testing, alongside the Zabbix agent (zabbix_agent2).

As we want to track individual applications, this is too coarse of an
approach. Instead, break out targets for individual application SBOM
generation and retain a target the entire module (sbom-mod).

The key change from this is that an `sbom.{json,xml}` file is no longer
generated and instead individual files per application and the module
are generated, following the naming pattern of
`sbom-{target}.{json,xml}`, where {target} is one of ms, za2, and zws,
for the mock_server, zabbix_agent2 and zabbix_web_service applications
repsectively, and simply `mod` for the entire module.

Additionally, provide `cyclonedx-gomod` with all Go build constraint
environment variables to mimic the actual build more closely.