template_app_systemd: change "service not running" prototype to "service failed"
It is absolutely common that systemd service are not running. The
prime example are timer activated services that enter the not running
state after they ran, waiting for the next timer activation.

The existing "Service not running" trigger prototype causes a lot of
noise due to false positives. Therefore, we replace this trigger
prototoype with one that explicitly warns about failed services,
because those need immediate attention.

Fixes ZBXNEXT-8767