.......... [DEV-2707] broke out more granular go SBOM targets
Previously, we generated a single SBOM for the entire module. However,
the module also carries other applications, alongside the Zabbix agent2.

As we want to track individual applications, this is too coarse of an
approach. Instead, break out targets for individual application SBOM
generation and retain a target for the entire module (sbom-mod).

The key change from this is that an `sbom.{json,xml}` file is no longer
generated and instead individual files per application and the module
are generated, following the naming pattern of
`sbom-{target}.{json,xml}`, where {target} is one of za2 or mod, for the
zabbix_agent2 application and the entire module repsectively. A
convenience target `sbom-apps` is also provided to generate SBOMs for
all applications.

Additionally, provide `cyclonedx-gomod` with all Go build constraint
environment variables to mimic the actual build more closely.

While here, declare non-artifact producing targets as such PHONY deps.