.......... [ZBXNEXT-8593] fixed integration test resource creation
Due to the mind numbing PHP Unit order of execution, data providers
execute before fixture setup and default data prep methods, meaning we
they can't reference results from those stages, for example, any
connectivity handles or data created as a fixture. The class setup
fixture can't report errors, only fail hard.

This is also before we fully initialize, where in our derived class, we
create a database session record to get a token for use in API requests.

These changes are in effect a crutch to work around this, by splitting
out the database resource creation into separate methods and keeping a
static lock identifying we've created these resources.

As these are called prior to the regular order in which the database
connecton is established seamlessly, we need to do this ourselves
explicitly.