.......... [ZBXNEXT-8593] fixed successful script manual input test case expected result messages
As the positive test simply echoes a string, optionally containing the
expandable macro, the resultant message ends in a newline, so add them
to the expected result messages. Do this in favor of calling echo with
the the `-n` flag (no newline) to keep it plain, simple and as expected.