....I..... [ZBX-23881] reordered compile checks in order of preference
cc and c99 are POSIX interfaces to a C-language compilation system and a
C99 standard compliant compiler system respectively. Although cc has
been marked LEGACY, both of these interfaces are more than commonly
found on modern day unices and its clones. The AC_PROG_CC macro
evaluates the list in the order specified, and with the previous order,
we always selected cc as the choice, because the test would always
succeed on platforms that provided it.