.......... [ZBX-23986] added a copy of GPLv2 as LICENSE
This is needed for Go license detection tools to determine the license
used for the module. As the root of the module is here, the COPYING file
at the root of the repository is not enough.

Things of note: the Golang license policy[1] recognizes a wide array of
filenames and their permutations to indicate the location of the license
text. The license file added is called LICENSE.

The Go Module Reference[2] mentions a Special Case[3] for the naming
choice of the file containg the license text. This mainly concerns
multiple module repositories and module authentication[4].

[1]: https://pkg.go.dev/license-policy
[2]: https://go.dev/ref/mod
[3]: https://go.dev/ref/mod#:~:text=Special%20case%20for%20LICENSE%20files
[4]: https://go.dev/ref/mod#authenticating