.........T [ZBXCTR-22] Refactor filters
This simple regex is about 2x as fast as the previous.

Arguably there's little point to verify the prefix/key follows the
specs for labels[1] or annotations[2], we're not inserting data, only
filtering on already validated data. Leave it up to the user to spell
the labels and annotations correctly.

[1] https://kubernetes.io/docs/concepts/overview/working-with-objects/labels/#syntax-and-character-set
[2] https://kubernetes.io/docs/concepts/overview/working-with-objects/annotations/#syntax-and-character-set