...G...... [ZBX-24531] fixed PKG in the mingw Makefile to reflect the updated path
In 85cbfdfa I updated the TARGETPKG variable to reflect the change of
the module name. Somehow, I missed the PKG variable, that's used to
avoid repetition, and as a result, the -X flags passed to the Go linker
had no effect, and the version message emitted by the agent didn't
display the build time information.