ZBX-18867: Change go env variables to be evaluated in makefile
Ensure that go variables are respected if defined outside the `go env`
environment. Historically the Zabbix go environment variables
`GOOS,GOARCH` have been evaluated based on `go env <var>`. This is the
default value that will be specified if no arguments are supplied.

This patch allows for these to be defined outside the go environment
and ensures that the variables are respected if passed in via the
commandline or via other scripts (needed for `cross-compiling on
platforms without another work around).

This has been discussed in a previously closed issue: ZBX-18867, it was
never discussed to remove the variables entirely as they are not needed.