....I..... [DEV-3893] added lookup of awk and a c++ compiler
The misc/images/svg_to_png.sh script depends on awk unconditionally.
We probably should look for it and not rely blindly on it just being
there. Also added a lookup for a c++ compiler, as we have C++ code that
gets built by it (src/libs/zbxsysinfo/win32/wmi.cpp).

Reordered the programs lookup in order from cheaper to more expensive.