....I..... [ZBX-23881] dropped compiler list from AC_PROG_CC call
cc and c99 are POSIX interfaces to a C-language compilation system and a
ISO/IEC 9899:1999 (aka C99) standard compliant compiler system
respectively. Although cc has been marked LEGACY, both of these
interfaces are more than common on modern day unices and clones.

The AC_PROG_CC macro evaluates the list in the order specified, and with
the list provided, we more often than not would use whatever `cc` in
$PATH would be instead of any of the other options.

Instead of doing any of this, just rely on the default behavior of the
macro, which is to use $CC from the environment if it's not empty, and
otherwise test for gcc, cc, followed by "other" compilers[1].

If the user wants to use a different compiler, $CC is the way.

[1]: https://www.gnu.org/software/autoconf/manual/autoconf-2.69/html_node/C-Compiler.html#:~:text=AC_PROG_CC