...G...PS. [ZBXNEXT-9338] changed flag param type of zbx_regexp_compile_ext
This parameter is a bitmask of PCRE2_* flags that gets passed to
pcre2_compile(3), which expects it to be an uint32_t, so change the type
wherever we're accepting the flags.

There are only 4 calls to zbx_regexp_compile_ext() in the entire tree,
and all of them pass the literal 0 as the value, so this ultimately
doesn't affect them.