...GI..PS. [ZBX-25140] defined ZBX_THREAD_LOCAL unconditionally
The server and proxy require thread-local storage since 6.4, when
preprocessing became multithreaded. The only circumstance where this was
an issue was Solaris 10 installations on SPARCs, where the system
compiler was a gcc3 pre-3.3.1.

Move the compiler test for recognizing the __thread keyword early in
configure and error immediately if the compiler fails the test.