...G...PS. [ZBX-25296] dropped overextended preproc test of LIBRESSL_VERSION_NUMBER
These tests were excessive, and effectively implied LibreSSL does not
implement the API beyond OpenSSL 1.0.1, which is not the case.

This results in a static definition of OPENSSL_cleanup() becoming
visible and conflicting with the prototype in openssl/crypto.h, thus
failing the build.

Theo Buehler (tb[at]openbsd[dot]org) committed this as a patch to the
ports tree a little over two years ago to work around this.

Praise to Theo for the patch!