...G...... [ZBX-25382] fixed NetBSD get_ifdata to unbreak builds on >= 8.2
The function used the kvm interface to retrieve network statistics. The
structure ifnet struct lost that member (if_data) in the 8.2 release
cycle, and we've been FTBFS ever since. Downstream this was resolved via
a patch in pkgsrc, originally submitted by otis[at]netbsd[dot]org for
5.0, at least 2 and a half years ago.

These changes are a modified version of that patch to fit the rest of
the code. In particular:

  * the switch..case has been dropped in favor of an if..else,
  * the validation of the if_name param accidentally nuked by the patch
    has been added back,
  * messages other than RT_IFINFO short-circuit the loop

Praise goes to otis@ for this! Thank you!