...G...PS. [DEV-4137] dropped loops from memory allocating wrappers
Though processors have become faster, and generally more processes get
scheduled in the same wall clock window, the loops are were a naive
attempt at getting memory in case the kernel finds some inbetween the
cycles, so drop them.

The other issue here is that the {c,m,re}alloc wrappers would behave
like glibc, and in the case of 0-sized allocations would bump that size
to at least 1 (or 1 member in the case of realloc's nmemb == 0). These
have been replaced with LOG_LEVEL_CRIT message emission notifying of a
bug, akin to how we do for when we re-use a pointer that already
contains a valid pointer obtained from a previous allocation.

The behavior has generally remained the same (sans the repeat attempts).