Added a special DYNAMIC_SEVERITY tag which allows overriding a trigger severity
The severity can be changed by raising it by a certain amount, lowering it, or setting a new
severity. When trigger that fires has the "DYNAMIC_SEVERITY" tag (which could be coming from
the template, host, trigger) the severity is set/adjusted according to the tag value.
Where "+N" raises the severity by N, "-N" lowers it by N, and just "N" (no plus or minus)
sets the severity to N. The "DYNAMIC_SEVERITY" tag can include macros so for example you can
have the ping trigger with the tag DYNAMIC_SEVERITY={$PING_SEVERITY} and set the macro to +1
in production hosts.